valentin comments

+++ crates/solver/src/main.rs
@@ -176,9 +176,9 @@ struct Arguments {
     #[structopt(long, env, default_value = "10")]
     paraswap_slippage_bps: u32,
 
-    /// The slippage tolerance we apply to the price quoted by Paraswap
+    /// The slippage tolerance we apply to the price quoted by zeroEx
     #[structopt(long, env, default_value = "10")]
-    zeroex_slippage_bps: u16,
+    zeroex_slippage_bps: u32,
 
     /// The authorization for the archer api.
     #[structopt(long, env)]
